lariat_agents/agent/event_payload/requirements.txt
attrs==21.4.0
boto3
botocore
certifi>=2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cramjam>=2.5.0
cryptography>=39.0.1
fastparquet>=0.8.1
fsspec>=2024.2.0
genson==1.2.2
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
numpy
packaging>=21.3
pandas>=2.0,<2.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyarrow>=14.0.1
pycparser==2.21
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.7.0
python-dateutil>=2.8.2
pytz>=2022.1
requests==2.27.1
responses==0.20.0
s3transfer
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
urllib3==1.26.9
Werkzeug>=2.2.3
xmltodict==0.13.0
zstd==1.5.2.5
s3fs>=2024.2.0
sqlparse==0.4.2
s3path==0.3.4
ruamel.yaml
ruamel.yaml.clib
boto3-type-annotations
datadog-api-client==2.11.0
python-dotenv==1.0.0
azure-core==1.26.0
azure-functions==1.12.0
azure-storage-blob==12.14.1
sqlglot
croniter>=2.0.3
pydantic>=1.10.2
pydantic_core>=2.10.1
python-magic>=0.4.27
python-snappy>=0.7.1
shapely>=2.0.4
pycountry
geopandas==0.14.4
google-cloud-storage==1.19.1
fastavro==1.9.4
